requirements.txt
# Core dependencies for Grounded Attention MVP

# PyTorch and Vision
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformers and related
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.7.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99

# Vision and multimodal
timm>=0.9.10
einops>=0.7.0
ftfy>=6.1.1

# Data processing
datasets>=2.15.0
Pillow>=10.0.0
opencv-python>=4.8.0
pycocotools>=2.0.7

# Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0

# NLP utilities
nltk>=3.8.1
spacy>=3.7.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Logging and experiment tracking
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0.0
omegaconf>=2.3.0

# Utilities
ipython>=8.18.0
jupyter>=1.0.0
black>=23.11.0
pytest>=7.4.0
